IGNITE-22000 Remove unused classes.
